+++ .classpath
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <classpath>
 	<classpathentry kind="src" path="src/main/java"/>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-17">
+	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER">
 		<attributes>
 			<attribute name="module" value="true"/>
 		</attributes>
@@ -13,25 +13,25 @@
 	</classpathentry>
 	<classpathentry kind="con" path="org.eclipse.jst.j2ee.internal.web.container"/>
 	<classpathentry kind="con" path="org.eclipse.jst.j2ee.internal.module.container"/>
-	<classpathentry kind="lib" path="C:/Users/valde/Desktop/connect-ong_clone/connect-ong/Modules/antlr-2.7.6.jar"/>
-	<classpathentry kind="lib" path="C:/Users/valde/Desktop/connect-ong_clone/connect-ong/Modules/antlr-2.7.7.jar"/>
-	<classpathentry kind="lib" path="C:/Users/valde/Desktop/connect-ong_clone/connect-ong/Modules/classmate-1.3.0.jar"/>
-	<classpathentry kind="lib" path="C:/Users/valde/Desktop/connect-ong_clone/connect-ong/Modules/commons-collections-3.1.jar"/>
-	<classpathentry kind="lib" path="C:/Users/valde/Desktop/connect-ong_clone/connect-ong/Modules/dom4j-1.6.1.jar"/>
-	<classpathentry kind="lib" path="C:/Users/valde/Desktop/connect-ong_clone/connect-ong/Modules/hibernate3.jar"/>
-	<classpathentry kind="lib" path="C:/Users/valde/Desktop/connect-ong_clone/connect-ong/Modules/hibernate-commons-annotations-5.0.1.Final.jar"/>
-	<classpathentry kind="lib" path="C:/Users/valde/Desktop/connect-ong_clone/connect-ong/Modules/hibernate-core-5.2.11.Final.jar"/>
-	<classpathentry kind="lib" path="C:/Users/valde/Desktop/connect-ong_clone/connect-ong/Modules/hibernate-jpa-2.0-api-1.0.0.Final.jar"/>
-	<classpathentry kind="lib" path="C:/Users/valde/Desktop/connect-ong_clone/connect-ong/Modules/javassist-3.12.0.GA.jar"/>
-	<classpathentry kind="lib" path="C:/Users/valde/Desktop/connect-ong_clone/connect-ong/Modules/javassist-3.20.0-GA.jar"/>
-	<classpathentry kind="lib" path="C:/Users/valde/Desktop/connect-ong_clone/connect-ong/Modules/jboss-logging-3.3.0.Final.jar"/>
-	<classpathentry kind="lib" path="C:/Users/valde/Desktop/connect-ong_clone/connect-ong/Modules/jta-1.1.jar"/>
-	<classpathentry kind="lib" path="C:/Users/valde/Desktop/connect-ong_clone/connect-ong/Modules/lucene-core-5.5.4.jar"/>
-	<classpathentry kind="lib" path="C:/Users/valde/Desktop/connect-ong_clone/connect-ong/Modules/lucene-queryparser-5.5.4.jar"/>
-	<classpathentry kind="lib" path="C:/Users/valde/Desktop/connect-ong_clone/connect-ong/Modules/mysql-connector-java-5.1.49.jar"/>
-	<classpathentry kind="lib" path="C:/Users/valde/Desktop/connect-ong_clone/connect-ong/Modules/mysql-connector-java-8.0.28.jar"/>
-	<classpathentry kind="lib" path="C:/Users/valde/Desktop/connect-ong_clone/connect-ong/Modules/postgresql-9.4-1201.jdbc41.jar"/>
-	<classpathentry kind="lib" path="C:/Users/valde/Desktop/connect-ong_clone/connect-ong/Modules/slf4j-api-1.6.1.jar"/>
-	<classpathentry kind="lib" path="C:/Users/valde/Desktop/connect-ong_clone/connect-ong/Modules/xml-apis-1.3.03.jar"/>
+	<classpathentry kind="lib" path="/Users/yago.pereira/Git/connect-ong/src/main/webapp/WEB-INF/lib/antlr-2.7.6.jar"/>
+	<classpathentry kind="lib" path="/Users/yago.pereira/Git/connect-ong/src/main/webapp/WEB-INF/lib/antlr-2.7.7.jar"/>
+	<classpathentry kind="lib" path="/Users/yago.pereira/Git/connect-ong/src/main/webapp/WEB-INF/lib/classmate-1.3.0.jar"/>
+	<classpathentry kind="lib" path="/Users/yago.pereira/Git/connect-ong/src/main/webapp/WEB-INF/lib/commons-collections-3.1.jar"/>
+	<classpathentry kind="lib" path="/Users/yago.pereira/Git/connect-ong/src/main/webapp/WEB-INF/lib/dom4j-1.6.1.jar"/>
+	<classpathentry kind="lib" path="/Users/yago.pereira/Git/connect-ong/src/main/webapp/WEB-INF/lib/hibernate-commons-annotations-5.0.1.Final.jar"/>
+	<classpathentry kind="lib" path="/Users/yago.pereira/Git/connect-ong/src/main/webapp/WEB-INF/lib/hibernate-core-5.2.11.Final.jar"/>
+	<classpathentry kind="lib" path="/Users/yago.pereira/Git/connect-ong/src/main/webapp/WEB-INF/lib/hibernate-jpa-2.0-api-1.0.0.Final.jar"/>
+	<classpathentry kind="lib" path="/Users/yago.pereira/Git/connect-ong/src/main/webapp/WEB-INF/lib/hibernate3.jar"/>
+	<classpathentry kind="lib" path="/Users/yago.pereira/Git/connect-ong/src/main/webapp/WEB-INF/lib/javassist-3.12.0.GA.jar"/>
+	<classpathentry kind="lib" path="/Users/yago.pereira/Git/connect-ong/src/main/webapp/WEB-INF/lib/javassist-3.20.0-GA.jar"/>
+	<classpathentry kind="lib" path="/Users/yago.pereira/Git/connect-ong/src/main/webapp/WEB-INF/lib/jboss-logging-3.3.0.Final.jar"/>
+	<classpathentry kind="lib" path="/Users/yago.pereira/Git/connect-ong/src/main/webapp/WEB-INF/lib/jta-1.1.jar"/>
+	<classpathentry kind="lib" path="/Users/yago.pereira/Git/connect-ong/src/main/webapp/WEB-INF/lib/lucene-core-5.5.4.jar"/>
+	<classpathentry kind="lib" path="/Users/yago.pereira/Git/connect-ong/src/main/webapp/WEB-INF/lib/lucene-queryparser-5.5.4.jar"/>
+	<classpathentry kind="lib" path="/Users/yago.pereira/Git/connect-ong/src/main/webapp/WEB-INF/lib/mysql-connector-java-5.1.49.jar"/>
+	<classpathentry kind="lib" path="/Users/yago.pereira/Git/connect-ong/src/main/webapp/WEB-INF/lib/mysql-connector-java-8.0.28.jar"/>
+	<classpathentry kind="lib" path="/Users/yago.pereira/Git/connect-ong/src/main/webapp/WEB-INF/lib/postgresql-9.4-1201.jdbc41.jar"/>
+	<classpathentry kind="lib" path="/Users/yago.pereira/Git/connect-ong/src/main/webapp/WEB-INF/lib/slf4j-api-1.6.1.jar"/>
+	<classpathentry kind="lib" path="/Users/yago.pereira/Git/connect-ong/src/main/webapp/WEB-INF/lib/xml-apis-1.3.03.jar"/>
 	<classpathentry kind="output" path="build/classes"/>
 </classpath>
