+++ .settings/org.eclipse.jst.j2ee.ejb.annotations.xdoclet.prefs
@@ -1,0 +1,5 @@
+XDOCLETBUILDERACTIVE=true
+XDOCLETHOME=
+XDOCLETUSEGLOBAL=true
+XDOCLETVERSION=1.2.3
+eclipse.preferences.version=1
