+++ src/main/java/hibernate.cfg.xml
@@ -3,7 +3,7 @@
 
 <hibernate-configuration>
 	<session-factory>
-		<property name="hibernate.connection.url">jdbc:mysql://localhost:3306/connect_ong</property>
+		<property name="hibernate.connection.url">jdbc:mysql://localhost:3306/connect_ong?useSSL=false</property>
 		<property name="hibernate.connection.username">root</property>
 		<property name="hibernate.connection.password"></property>
 		<property name="hibernate.connection.driver_class">com.mysql.jdbc.Driver</property>
@@ -23,6 +23,3 @@
 		-->
 	</session-factory>
 </hibernate-configuration>
-<!--  
-		
--->
