+++ src/main/java/model/Ong.java
@@ -56,6 +56,5 @@ public class Ong {
 	}
 	
 	
-	
 
 }
