+++ src/main/java/teste/teste.java
@@ -1,0 +1,21 @@
+package teste;
+
+import dao.DaoTelefone;
+import model.Telefone;
+
+public class teste {
+
+	public static void main(String[] args) {
+	
+		DaoTelefone daotel = new DaoTelefone();
+		Telefone tel = new Telefone("32816655", "75", "55");
+				
+		daotel.save(tel);
+		System.out.println(daotel.retrieveById(1).getNumero());
+		System.out.println("Teste");
+		
+
+		
+	}
+
+}
