+++ src/main/webapp/index.jsp
@@ -1,5 +1,5 @@
 <%@page import="java.util.List"%>
-<%@page import="controller.DaoTelefone"%>
+<%@page import="dao.DaoTelefone"%>
 <%@page import="model.Telefone"%>
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
@@ -12,14 +12,11 @@
 </head>
 <body>
 	<% 
-		DaoTelefone daotel = new DaoTelefone();
+		DaoTelefone daoTelefone = new DaoTelefone();
 		Telefone telefone = new Telefone("8888-7777", "75", "+001");
-		List<Telefone> telDb = daotel.retrieveAll();
-		
-		for(Telefone tel : telDb) {
 	%>
-			<h2>Hello World From Database <% out.print(tel.getNumero()); %></h2>
 			
-		<%}%>
+		
+<h2>Hello World From Database <% out.print(telefone.getNumero()); %> </h2>
 </body>
 </html>
